social_media_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
django-filter==25.1
drf-nested-routers==0.95.0

